+++ app/views/cclow/cclow/_list_item.html.erb
@@ -2,6 +2,6 @@
   <h2><%= title %></h2>
 
   <p>
-    <%= description %>
+    <%= description&.html_safe %>
   </p>
 </div>
