+++ app/views/cclow/geography/legislations/show.html.erb
