+++ config/routes.rb
@@ -21,10 +21,10 @@ Rails.application.routes.draw do
     root to: 'home#index'
 
     resources :geographies, only: [:show] do
-      resources :laws, controller: :legislations, only: [:show, :index], defaults: { scope: :laws }
-      resources :policies, controller: :legislations, only: [:show, :index], defaults: { scope: :policies }
-      resources :litigation_cases, only: [:show, :index]
-      resources :climate_targets, only: [:show, :index]
+      resources :laws, controller: 'geography/legislations', only: [:show, :index], defaults: { scope: :laws }
+      resources :policies, controller: 'geography/legislations', only: [:show, :index], defaults: { scope: :policies }
+      resources :litigation_cases, controller: 'geography/litigation_cases', only: [:show, :index]
+      resources :climate_targets, controller: 'geography/climate_targets', only: [:show, :index]
     end
   end
